refactor stats macro
