+++ varnish-macros/src/lib.rs
@@ -2,8 +2,8 @@
 // #![allow(warnings)]
 
 use errors::Errors;
-use quote::{format_ident, quote};
-use syn::{parse_macro_input, Data, DeriveInput, Fields, ItemMod, Type};
+use quote::format_ident;
+use syn::{parse_macro_input, DeriveInput, ItemMod};
 use {proc_macro as pm, proc_macro2 as pm2};
 
 use crate::gen_docs::generate_docs;
@@ -20,6 +20,7 @@ mod names;
 mod parser;
 mod parser_args;
 mod parser_utils;
+mod stats;
 
 pub(crate) type ProcResult<T> = Result<T, Errors>;
 
@@ -76,181 +77,12 @@ pub fn stats(_args: pm::TokenStream, input: pm::TokenStream) -> pm::TokenStream 
     let name = &input.ident;
     let vis = &input.vis;
 
-    let fields = match &input.data {
-        Data::Struct(data) => match &data.fields {
-            Fields::Named(fields) => &fields.named,
-            _ => panic!("Only named fields are supported"),
-        },
-        _ => panic!("Only structs are supported"),
-    };
+    let fields = stats::get_struct_fields(&input.data);
+    stats::validate_fields(fields);
 
-    // Ensure all fields are AtomicU64, for now
-    for field in fields {
-        match &field.ty {
-            Type::Path(path) => {
-                let is_atomic_u64 = path
-                    .path
-                    .segments
-                    .last()
-                    .is_some_and(|seg| seg.ident == "AtomicU64");
-
-                if !is_atomic_u64 {
-                    let field_name = field.ident.as_ref().unwrap();
-                    panic!("Field {field_name} must be of type AtomicU64");
-                }
-            }
-            _ => panic!("Field types must be AtomicU64"),
-        }
-    }
-
-    let original_fields = fields.iter().map(|f| {
-        let name = &f.ident;
-        let vis = &f.vis;
-        quote! { #vis #name: std::sync::atomic::AtomicU64 }
-    });
-
-    let metrics = fields.iter().map(|field| {
-        let field_name = field.ident.as_ref().unwrap().to_string();
-        
-        // Determine counter type from attributes
-        let counter_type = if field.attrs.iter().any(|attr| attr.path().is_ident("counter")) {
-            "counter"
-        } else if field.attrs.iter().any(|attr| attr.path().is_ident("gauge")) {
-            "gauge"
-        } else {
-            panic!("Field {field_name} must have either #[counter] or #[gauge] attribute")
-        };
-
-        // Extract documentation from doc comments
-        let mut doc_lines = field.attrs.iter()
-            .filter(|attr| attr.path().is_ident("doc"))
-            .filter_map(|attr| {
-                let syn::Meta::NameValue(meta) = &attr.meta else { return None };
-                let syn::Expr::Lit(expr) = &meta.value else { return None };
-                let syn::Lit::Str(lit) = &expr.lit else { return None };
-                Some(lit.value())
-            })
-            .filter(|s| !s.is_empty());
-
-        let oneliner = doc_lines.next().unwrap_or_default();
-        let docs = doc_lines.next().unwrap_or_default();
-
-        // Parse optional attributes
-        let mut level = String::from("info");
-        let mut format = String::from("integer");
-        
-        if let Some(attrs) = field.attrs.iter()
-            .find(|attr| attr.path().is_ident(counter_type)) {
-            
-            let _ = attrs.parse_nested_meta(|meta| {
-                match meta.path.get_ident().map(ToString::to_string).as_deref() {
-                    Some("level") => {
-                        level = meta.value()?.parse::<syn::LitStr>()?.value();
-                    }
-                    Some("format") => {
-                        format = meta.value()?.parse::<syn::LitStr>()?.value();
-                        assert!(["integer", "bitmap", "duration", "bytes"].contains(&format.as_str()), "Invalid format value for field {field_name}. Must be one of: integer, bitmap, duration, bytes")
-                    }
-                    _ => {}
-                }
-                Ok(())
-            });
-        }
-
-        quote! {
-            VscMetricDef {
-                name: #field_name,
-                counter_type: #counter_type,
-                ctype: "uint64_t",
-                level: #level,
-                oneliner: #oneliner,
-                format: #format,
-                docs: #docs,
-            }
-        }
-    });
-
+    let original_fields = stats::generate_field_definitions(fields);
+    let metrics = stats::generate_metrics(fields);
     let name_inner = format_ident!("{}Inner", name);
 
-    quote! {
-        use varnish::ffi::vsc_seg;
-        use varnish::ffi::{VRT_VSC_Alloc, VRT_VSC_Destroy};
-        use varnish::vsc_types::{VscMetricDef, VscCounterStruct};
-        use std::ops::{Deref, DerefMut};
-        use std::ffi::CString;
-
-        #[repr(C)]
-        #[derive(Debug)]
-        #vis struct #name_inner {
-            #(#original_fields,)*
-        }
-
-        // Wrapper struct to hold the VSC segment and the inner counter struct
-        #vis struct #name {
-            value: *mut #name_inner,
-            vsc_seg: *mut vsc_seg,
-            name: CString,
-        }
-
-        impl Deref for #name {
-            type Target = #name_inner;
-
-            fn deref(&self) -> &Self::Target {
-                unsafe { &*self.value }
-            }
-        }
-
-        impl DerefMut for #name {
-            fn deref_mut(&mut self) -> &mut Self::Target {
-                unsafe { &mut *self.value }
-            }
-        }
-
-        impl varnish::vsc_types::VscCounterStruct for #name {
-            fn set_vsc_seg(&mut self, seg: *mut vsc_seg) {
-                self.vsc_seg = seg;
-            }
-
-            // Introspect the struct to get the metric definitions
-            fn get_struct_metrics() -> Vec<varnish::vsc_types::VscMetricDef<'static>> {
-                vec![
-                    #(#metrics),*
-                ]
-            }
-
-            fn new(module_name: &str, module_prefix: &str) -> #name {
-                let mut vsc_seg = std::ptr::null_mut();
-                let name = CString::new(module_name).unwrap();
-                let format = CString::new(module_prefix).unwrap();
-
-                let json = Self::build_json(module_name);
-
-                let value = unsafe {
-                    VRT_VSC_Alloc(
-                        std::ptr::null_mut(),
-                        &mut vsc_seg,
-                        name.as_ptr(),
-                        size_of::<#name_inner>(),
-                        json.as_ptr(),
-                        json.len(),
-                        format.as_ptr(),
-                        std::ptr::null_mut()
-                    ) as *mut #name_inner
-                };
-
-                return #name {
-                    value,
-                    vsc_seg,
-                    name,
-                }
-            }
-
-            fn drop(&mut self) {
-                unsafe {
-                    VRT_VSC_Destroy(self.name.as_ptr(), self.vsc_seg);
-                }
-            }
-        }
-    }
-    .into()
+    stats::generate_output(name, &name_inner, vis, original_fields, metrics).into()
 }
